requirements.txt
Django==1.6.1
MySQL-python==1.2.5
Pillow==2.5.3
South==0.8.4
django-auth-ldap==1.1.7
django-moderation==0.3.6
django-piwik==0.1
django-qrcode==0.3
django-simple-captcha==0.4.2
gunicorn==18.0
piwikapi==0.3
python-ldap==2.4.13
simplejson==3.3.1
six==1.7.3
wsgiref==0.1.2
